requirements.txt
requests==2.32.3
pyyaml==6.0.2
pytz==2024.1
google-generativeai==0.7.2
python-dateutil==2.9.0.post0
tenacity==8.3.0
